crimson/tools/store-nbd: break into multiple files
