+++ src/crimson/tools/CMakeLists.txt
@@ -1,5 +1,7 @@
 add_executable(crimson-store-nbd
-	store-nbd.cc
+	store_nbd/store-nbd.cc
+	store_nbd/tm_driver.cc
+	store_nbd/block_driver.cc
   )
 target_link_libraries(crimson-store-nbd
   crimson-seastore)
